requirements.txt
argon2-cffi==25.1.0
fastapi==0.136.0
loguru==0.7.3
pydantic==2.13.3
pydantic-settings==2.14.0
PyJWT==2.12.1
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2026.1.post1
tortoise-orm[accel,asyncpg]==1.1.7
uvicorn[standard]==0.44.0
python-escpos[all]==3.1
